requirements.txt
accelerate==1.9.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
ase==3.25.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
beartype==0.21.0
bitsandbytes==0.46.1
captum==0.8.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
CoLT5-attention==0.11.1
contourpy==1.3.2
cycler==0.12.1
DateTime==6.0
descriptastorus==2.8.0
e3nn==0.5.6
echo_logger==0.1.6
einops==0.8.1
einops-exts==0.0.4
einx==0.3.0
et_xmlfile==2.0.0
faiss-cpu==1.12.0
filelock==3.13.1
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
hyper-connections==0.2.1
idna==3.10
jcamp==1.2.2
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.9
lmdb==1.7.2
local-attention==1.11.2
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openpyxl==3.1.5
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
packaging==25.0
paddlepaddle-gpu==2.6.2
pandas==2.3.1
pandas_flavor==0.7.0
periodictable==2.0.2
pgl==2.2.6
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rdkit==2025.3.3
regex==2024.11.6
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.33.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.6.0+cu118
torch-geometric==2.6.1
torch_cluster==1.6.3+pt26cu118
torch_scatter==2.1.2+pt26cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.53.2
Tree==0.2.4
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vector-quantize-pytorch==1.22.18
wandb==0.21.0
xarray==2025.6.1
yarl==1.20.1
zetascale==2.8.6
zope.interface==8.2